Move ActorID type into its own header [llm]
As a rule, in headers, we forward declare structs instead of including
the .h file, but we can't really do that for typedefs that refer to
basic types, such as ActorID.

ActorID being typedef'd in actor.h was forcing headers to include it.
This removes that necessity, and now they can just include actor_id.h.